requirements.txt
Django>=5.2
djangorestframework>=3.14.0
coreapi>=2.3.3
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
drf-spectacular>=0.27.1 
rest_framework_simplejwt